Writing the code 0.2

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,16 +7,16 @@
 int main(void)
 {
 	int i = 1;
-	long int num1 = 0;
-	long int num2 = 1;
-	long int sumPrevTwo = 0;
+	long long int num1 = 0;
+	long long int num2 = 1;
+	long long int sumPrevTwo = 0;
 
 	while (i <= 98)
 	{
 		if (i < 100)
-			printf("%ld%s", num1 + num2, ", ");
+			printf("%lld%s", num1 + num2, ", ");
 		else
-			printf("%ld", num1 + num2);
+			printf("%lld", num1 + num2);
 		sumPrevTwo = num1 + num2;
 		num1 = num2;
 		num2 = sumPrevTwo;
